0x15-file_io: 0-read_textfile.c file updated 7th

+++ 0x15-file_io/0-read_textfile.c
@@ -15,7 +15,7 @@
 */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-int file = open(filename, O_RDONLY);
+ssize_t file = open(filename, O_RDONLY);
 char *buff;
 ssize_t bytesR = read(file, buff, letters);
 ssize_t bytesW = write(STDOUT_FILENO, buff, bytesR);
